+++ src/region.h
@@ -46,7 +46,7 @@ public:
 	int rebuild(const split_interval_map *_mmap);
 	int print(int index) const;
 	bool is_allelic() const;
-	int assign_as_cov(int _ave, int _dev, int _max);
+	int assign_as_cov(double _ave, double _dev, double _max);
 	bool operator<(const region &x) const;
 
 private:
